% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIMGtopanel.R
\name{addIMGtopanel}
\alias{addIMGtopanel}
\title{Add images to a ggplot facet plot}
\usage{
addIMGtopanel(p1, img.list, pal = colorRamps::matlab.like(100))
}
\arguments{
\item{p1}{A ggplot object with panels, i.e..}

\item{img.list}{A list of images for each center. Image 1 goes with panel 1, etc. Panels are numbered across rows. Image are recycled if the list is shorter than the number of panels.}

\item{pal}{The color ramp for the images. This will determine what the images look like.}
}
\description{
This adds images in a image list to the panels in a ggplot object.
}
\details{

}
