% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat.R
\name{desat}
\alias{desat}
\title{Desaturate colors}
\usage{
desat(cols, sat = 0.5)
}
\arguments{
\item{cols}{a vector of colors}

\item{sat}{saturation level}
}
\value{
A list of desaturated colors.
}
\description{
Function for desaturating colors by specified proportion. Reference:
https://stackoverflow.com/questions/26314701/r-reducing-colour-saturation-of-a-colour-palette
}
\details{
See the \strong{colorspace} library for examples of \code{hsv()} colorspace manipulations and
\strong{RColorBrewer} library for some example colors
}
