% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat}
\alias{distMat}
\alias{corMat,}
\alias{`distMat`}
\alias{returns}
\alias{`stats::dist(X,}
\alias{diag=TRUE,}
\alias{upper=TRUE)`.}
\alias{`corMat`}
\alias{a}
\alias{distance}
\alias{matrix}
\alias{from}
\alias{the}
\alias{correlation}
\alias{matrix.}
\title{Distance matrices}
\usage{
distMat(X, remove.mean = FALSE)
}
\arguments{
\item{X}{A matrix}

\item{remove.mean}{If TRUE, will remove the mean from each row of X}
}
\description{
Distance matrices
}
