% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Download data from an ERDDAP server}
\usage{
getdata(
  id,
  pars = NULL,
  lat = c(7, 13),
  lon = c(72, 78),
  date = NULL,
  altitude = 10,
  alt.name = "altitude",
  eserver = "http://apdrc.soest.hawaii.edu/erddap",
  datadir = "data"
)
}
\arguments{
\item{id}{The id of the data. Like hawaii_soest_26fe_f2c9_ed3b. In quotes.}

\item{pars}{Name of the parameter. Like sst, e.g. In quotes.}

\item{lat}{Min-max latitude}

\item{lon}{Min-max longitude}

\item{date}{To specify a date range, pass this in like c("2020-01-01", "2020-02-01"). Leave off to download the whole data range.}

\item{altitude}{Optional. Some data types need this. Ignore if not. Look on the data page. time, latitude and longitude are always there. If altitude is also there, then put that in.}

\item{alt.name}{If altitude is present, what is it's name.}

\item{eserver}{The url to the erddap server}

\item{datadir}{The folder where to save the data.}
}
\value{
The function saves a csv file and returns the data list invisibly.
}
\description{
Download data using eserver name
}
\details{
Go to the server to see what the pars names are and the format. The names for
things are unique and can't be predicted.
e.g. http://apdrc.soest.hawaii.edu/erddap/griddap/hawaii_soest_26fe_f2c9_ed3b.html
The 2 main servers I use are https://coastwatch.pfeg.noaa.gov/erddap and http://apdrc.soest.hawaii.edu/erddap
Search from there to find the data you want.

Common ids on apdrc.soest.hawaii.edu server
\itemize{
\item ERA5 1950-1978 hawaii_soest_26fe_f2c9_ed3b
\item ERA5 1979-2021 hawaii_soest_d124_2bb9_c935
\item NOAA OISST AVHRR daily hawaii_soest_afc8_9785_907e; add 0.125 to lat and lon
\item NOAA OISST AVHRR daily 2016+ hawaii_soest_330b_094e_ca45
}
}
