% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kheatmap.R
\name{kheatmap}
\alias{kheatmap}
\title{Heat map function}
\usage{
kheatmap(
  n_K,
  X_norm,
  datalist,
  main.n = 5,
  sub.n = n_K,
  iter.max = 25,
  nstart = 100,
  plotit = TRUE,
  ramp = "var",
  ramp.type = 2,
  dend.type = "original"
)
}
\arguments{
\item{n_K}{Number of K cluster to use}

\item{X_norm}{The normed centers as a matrix, each row is a centroid image}

\item{datalist}{The list with dat, dat.clean and pos.loc.}

\item{main.n}{The number of main clusters. These will have different colors.}

\item{sub.n}{The number of sub levels. This will have different shades.}

\item{iter.max}{Passed to the kmeans function.}

\item{nstart}{Passed to the kmeans function.}

\item{plotit}{If TRUE, plot the heat map.}
}
\value{
A list with two types of objects: a heatmap and a dendogram.
}
\description{
Bespoke function for Quan Seminar talk.
}
\details{
\strong{Notes}: Probably too customized and prelimanary to keep.
}
