% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myheatmap.R
\name{myheatmap}
\alias{myheatmap}
\title{Plot the heat map}
\usage{
myheatmap(
  dend,
  kobject,
  main.n = 3,
  sub.n = 7,
  ramp = c("mean", "var"),
  ramp.type = 1,
  plotit = TRUE
)
}
\arguments{
\item{dend}{Dendogram object}

\item{kobject}{Object out of \code{kmeans()}}

\item{main.n}{The number of main clusters. These will have different colors.}

\item{sub.n}{The number of sub levels. This will have different shades.}

\item{ramp}{How to create the shading.}

\item{ramp.type}{Type of ramping of the shades.}

\item{plotit}{If TRUE, plot the heat map.}
}
\value{
A ggplot object.
}
\description{
Another bespoke function for Quan Seminar talk.
}
\details{
\strong{Notes}: Probably too customized and prelimanary to keep.
}
