% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCSV.R
\name{processCSV}
\alias{processCSV}
\title{Read in image csv}
\usage{
processCSV(file, lat_range, long_range, has.alt = FALSE)
}
\arguments{
\item{file}{Name of the csv file}

\item{lat_range}{What range to subset}

\item{long_range}{What range to subset}

\item{has.alt}{If TRUE, then remove 2nd column}

\item{aspect_ratio}{c(width, height). This is \code{c(length(unique(lons)), length(unique(lats)))}}
}
\value{
The function returns \code{dat.wide} which is the original data (in the specified lat/lon box)
where each row is a date and each column is a pixel in the image grid.
\code{dat.wide} may have NAs (say from land if working with ocean
data). \code{dat.clean} is the data with NA pixels (i.e. land) removed.
\code{pos.loc} is the location of the
non-NA pixels (columns in \code{dat.wide}) so that the image can be reconstructed after
k clustering is performed.
}
\description{
processCSV will read in a csv file of SST and return the matrices needed for various ML tasks.
}
\details{
\code{stats::kmeans()} wants a matrix where each row is a sample of your data. So we want each
row of our data matrix to be a date and the columns will be the pixels in the image.
}
