% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearTable.R
\name{yearTable}
\alias{yearTable}
\title{Plot images by years and month}
\usage{
yearTable(
  X2,
  method = c("kmeans", "hclust.complete", "hclust.ward", "dbscan"),
  K = 3,
  iter.max = 25,
  nstart = 100,
  eps = 7,
  dist.type = c("euclidian", "correlation", "euclidian2", "canberra")
)
}
\arguments{
\item{method}{What clustering function to use.}

\item{K}{Number of clusters}

\item{iter.max}{Passed to the kmeans function.}

\item{nstart}{Passed to the kmeans function.}

\item{dist.type}{Distance function to use in the clustering.}

\item{plotit}{If TRUE, plot the heat map.}
}
\value{
A list with two types of objects: a heatmap and a dendogram.
}
\description{
Bespoke function for Quan Seminar talk. Month on y-axis and year on x-axis. Each cell is a color-coded image type (e.g. upwelling or not)
}
\details{
\strong{Notes}: Probably too customized and prelimanary to keep.
}
